requirements.txt
SmartApi-Python>=1.3.2
pandas>=2.0.0
numpy>=1.24.0
ta
scikit-learn>=1.3.0
python-dotenv>=1.0.0
loguru>=0.7.0
fastapi>=0.103.0
uvicorn>=0.23.0
requests>=2.31.0
websocket-client>=1.6.1
pydantic>=2.4.0
pyotp>=2.9.0
